Hi,
I have made the changes as discussed, the code is attached to the
bugzilla bug concerning part of the issues that were found.
A few notes:

1. There is a ParserException raised on TreeBlock. Tough there is a
TreeBlockParser, most of the important parsing was (and still is!)
being made on TreeBlock. I would imagine that this is not the best
design, but I did not change it.
2. I made some test cases. Also included.
3. I don't mind producing some documentation, in case you accept the code.
4. I noticed a few minor bugs more (like eating spaces in the names of
nodes). But they are really minor.
5. The API was changed, but I suppose not many people were parsing
trees. If there were people parsing trees most probably the bug on not
being able to process trees that are not binary would have been
detected as it is pretty major.

Tiago



git-svn-id: http://code.open-bio.org/repos/biojava/biojava-live/trunk@7269 7c6358e6-4a41-0410-a743-a5b2a554c398