Updating changes for FastaReader to support marking the current index position in the stream to allow for sequence proxy in a file. The new features haven't been tested yet.
Added in support for a SequenceParserInterface to allow support for different ways a sequence may be stored in a file mainly the challenge of CR or LF may or may not be in the stream. The seaquence could also have header information with position etc. 

Put in test cases for FastaReader

Created FastaWriter class



git-svn-id: http://code.open-bio.org/repos/biojava/biojava-live/trunk@7406 7c6358e6-4a41-0410-a743-a5b2a554c398