Added a test package, org.biojava.bio.structure.align.benchmark.
These classes are in src/test with the intention that they will eventually provide a way to benchmark various StructureAlignment algorithms. For instance, running the unit tests would check that CE continues to provide alignments which match a benchmark to some threshold. This type of test is useful for large changes to the algorithm.

For now, the package does not run as a unit test, but merely outputs various statistics for CE alignments. See class AlignBenchmark for main().

git-svn-id: http://code.open-bio.org/repos/biojava/biojava-live/trunk@7613 7c6358e6-4a41-0410-a743-a5b2a554c398