Introduce a new Bond class
The previous Bond class was barely used, so this completely replaces it.

Bonds keep track of two atoms, their type (using BondType), and their
order.

Also included are a few utility methods, like those for getting
distances between the two atoms and getting the other atom in a bond,
given the first one.