Allow custom parameters in structure alignment CLIs.
The CLIs for most structure alignment algorithms derive from
AbstractUserArgumentParser, but there wasn't a good mechanism to
introduce new arguments, since all parameters are stored into a
bean.

Without refactoring everything, this makes the parameter specification
a bit more consistent. The standard parameters are still stored in
a StartupParameters bean, but this is extended by each class to store
the class-specific parameters. Help strings are taken from a
ConfigStrucAligParams subclass, which was previously only used for the
GUI parameters. So the GUI and CLI parameter lists should now match.

The procedure for adding a new CLI implementation is documented in
AbstractUserArgumentParser.

Methods converted to the new system:
* Ce
* CeCP
* FatCat
* SmithWaterman3D

Other AbstractUserArgumentParser implementations downstream of BioJava
will need to be updated to work with the new system.