Remove and replace StructureAlignmentException
It was only used for problems loading the Atom arrays and to check
consistency in some parts of the calculations, but the usage was not
clearly defined. 
The exceptions have been replaced by NullPointerException and
IllegalStateException, respectively, and since they are Java runtime
exceptions they do not need to be thrown. 
Because there was no need to catch these exceptions, so they needed to
be thrown always, the change does not affect the behavior of the code,
but simplifies it.