Adding ZipChemCompProvider: pulls ChemComp files from a zip archive or download. Adding ZipAppendUtil: manages updating the zip archive with new additions.
DownloadChemCompProvider: updated to add a check if the requested ChemComp
was found or threw FileNotFoundException.  downloadChemCompRecord() returns
true if it was successful.  getChemComp now checks if the download was
successful so that it won't throw an IOException if no chemcomp exists.

TestChemCompProvider unit test checks creation/addition with ZipChemCompProvider and includes 'bad' residues that would cause DownloadChemCompProvider to fail.