Fix #480. Correct spelling of _symmetry.space_group_name_H-M in CIF output
CIF categories are represented as a simple bean, typically extending AbstractBean.
By default, all fields from the bean are taken as the CIF labels. Fields may be omitted
by annotating them as @IgnoreField. The CIF label for a field may be changed
by defining a function, `static Map<String,String> getCIFLabelMap();`,
mapping from the field's name to the correct label.

This commit also changes the MMCIFFileTools.toMMCIF method to be generic,
so that we enforce at compile time that all loop rows must be of the same type.