Add generics to org.biojava.nbio.core.search.io package (#354)
This makes the BLAST parser typesafe with respect to the sequences it
produces. Most classes in the package are now parameterized according to
the Sequence they act on. This does have the downside that one should
know prior to parsing what type of sequence will be returned. Otherwise
one will get type safety warnings. This may be undesirable for the
common case of parsing unknown text, but at least it exposes the potential
problem to the user.