Fix CE-Symm bug relating to short subunits.
[rcsb/symmetry#118](https://github.com/rcsb/symmetry/issues/118)

The problem was actually downstream from CE-Symm in the QuatSymmDetector
code used for assigning a symmetry group. The code was re-aligning
subunits using CE. This fails for short alignments of <15 residues
(two AFP with default settings). Thus no atoms get aligned and an
exception was thrown whenever an output format required information
about the symmetry group.

This fix re-uses the original CE-Symm alignment for the symmetry group
detection. This might change CE-Symm and QuatSymm results for proteins
with small subunits or borderline scores. 

Changes:
- New `SubunitClustererMethod.MANUAL` enum value
- Improve `SymmetryTools.getQuaternarySymmetry`